Added mail branch. I made this due to issues I had with the emailBranch. Delete the earlier emailBranch.
This is due to the gitIgnore change added to development.
Added the SendEmail class in Services.
This takes in a subject, message, and target address.
Call sendMail to send the email.

Note you will have to download the JavaMail package for this to work
I'll provide the jar for that over discord.